Link to KSP install directory
Since the path is so long and contains spaces, it's a bit of a pain to
keep trying to find and cd to that directory. This will make it easier
for scripts to have access to things that are stored in the KSP root
like the save games, CKAN data, modules, and other configs.